feat: expand to 50 countries via generic Box Office Mojo provider
Refactor the US-only provider into a generic BoxOfficeMojoProvider that
supports all Box Office Mojo international markets via the ?area= URL
parameter. The same HTML parsing works for all markets.

50 countries now supported including: US, UK, Germany, France (BOM + AlloCiné),
Italy, Spain, Australia, Brazil, Mexico, South Korea, Japan, India, China,
and 37 more.

Architecture:
- BoxOfficeMojoProvider: generic provider parameterized by area code
- USProvider: thin subclass (area=None for domestic)
- FranceProvider: AlloCiné (admissions-based, unchanged)
- Dynamic UI dropdown populated from provider registry
